api/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==4.2.7
django-debug-toolbar==4.2.0
executing==2.0.1
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.10
gunicorn
djangorestframework
django-cors-headers

